Branding, trademark changes
Don't advertise the web app as the 'Fedora Update System', in our
deployment it is used for our own distribution, not for Fedora.

Also, we're not really a "community maintained site", so we shouldn't
say it, and we certainly shouldn't confuse people by waiving Red Hat's
responsibility on our site. (I mean, they already have none :)

There is still the problem that the Fedora logo is used, but there's not
much that can be done here (I don't want to bundle the Network Box logo
in the Bodhi package), so we'll make symlinks at %install time in the
spec file.