Use our Koji tags and repo names for our releases
This is a big change.

First, our versions are not numeric like in Fedora, it's 5.0, 5.1,... So
we need to change the way this is handled.

Also, if we're changing the way Bodhi guesses Koji tags and repo names,
then we might as well remove all the various conditions on pending
releases (we don't do that) and EPEL special casing.

It makes the patch bigger, but the overall code becomes simpler.