Updates Policy: Don't enforce a minimum time in testing
Our policy is much simpler: a proventester must ack the update.

Enforcing a minimum number of days in testing just complicates the
process, and doesn't bring us any benefit. Updates will just stay in
testing until they are acked, and will be moved to stable as soon as
they are.