Mash: Fix for our releases/repos/tags setup
Mash assumes that tags are called under the following convention:
    $release-testing
    $release-updates

So it splits them on the last '-' and takes the first part, then decides
that's what it needs.

In our case, this is problematic because the tags are named as follow:
    nb5.0-free
    nb5.0-free-testing

So splitting as it does, we end up with something wrongly inconsistent.