Updates Policy: All provenpackagers and releng can push to all packages
We don't use Fedora's PackageDB. Users authenticate against our
internal LDAP, but the package ACL system is set to 'dummy'.

So let's hardcode the group values we use.

And because we don't use PackageDB, I took the liberty to simplify the
API for this function, and remove useless code which was just setting up
the arguments related to PackageDB.