Updates Policy: Prevent anonymous karma
Only identified developers and QA people will be giving
karma. Ever.

We are not an open community creating a general-purpose
distribution, we are a security company creating a
specialized OS for our products.

So anonymous karma just makes no sense in our case.

Also this removes the dependency on TGCaptcha2, which is
always nice, as it's one less package I have to maintain. :)