Fix changelog handling
The RPM API does not seem to always return a list for the changelog
timestamp(s).

Bodhi has some code trying to identify that and treat it accordingly.

However, it was based on the identified cases where this happened (there
is only one changelog text) instead of just being based on whether or
not the changelog was a list.

It seems that this behaviour in the RPM Python API, as I'm now seeing
cases where Bodhi thinks the timestamp is not in a list, when it
actually is.

This commit fixes this.