Reset karma to 0 on edit
When testers submit karma to an update, it shouldn't be kept if the
update is edited and its builds are changed.

However, we keep the history of comments added, because that is
important.

Also, if no builds are changed (e.g just the update notes were
modified), then there's no point in resetting the karma to 0.

https://fedorahosted.org/bodhi/ticket/388
https://fedorahosted.org/fesco/ticket/1238