Don't depend on fedmsg
This is all pretty much optional in Bodhi at the moment, and Bodhi is
just emitting messages anyway.

But if we don't remove it all, then we need to package fedmsg and all
its dependencies (moksha, ...), and we'll have to set up the connection
endpoint which Bodhi tries to connect to,...

That's just not worth the hassle for a few messages which we won't make
anything of anyway.

That being said, removing all traces of fedmsg leads to a pretty
extensive patch, which is really hard to merge. (I know, I tried :)

This little hack turns out to be much simpler than the two other
alternatives.

We'll revisit when other parts of the Fedora build infrastructure are
doing cool stuff with fedmsg and we want to take advantage of it.