updateinfo: Prefix the file name with its hash
This is similar to what is done with other Yum repository metadata.

It is not just about consistency though, it fixes real issues.

For example, at work, we use the Amazon CDN to implement mirroring of
our Yum repositories. All files are cached on the CDN for the default
duration (i think it's 24 hours), except repomd.xml which is never
cached.

This works great for RPMS (they never change once published) and all
the metadata files (their names are prefixed with their hash, and as
such the CDN sees them as different files).

However, we keep running into trouble with updateinfo.xml.gz: the CDN
keeps distributing the old file even though a new one was generated.

Because the updateinfo metadata can be quite big, we'd rather prefix
its name with a hash (like other Yum repository metadata), and let the
CDN handle it, rather than telling the CDN to never cache it, which
would put more load on our primary repository server.